requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
setuptools
pymongo[srv]==3.12
scikit-learn
scipy
pyaml
mlflow
dagshub
fastapi
uvicorn
google-cloud-storage